implemented a new way to handle command buffers on renderer threads
I'm using the same N+1 strategy as before, but for secondary cmd buffers
this time. It's implemented in the CmdBufferRing struct. This allows
renderers to fully manage their command buffers.
Had to get rid of TLS because it was randomly mixing pools from different
threads, so we're back to back to one pool per renderer object.
This gets rid of most of the allocator churn and improves performance,
mostly on amdvlk