flip the Z coordinate in the vertex shader
this should solve models being flipped. but it's a workaround, and there
must be a better way that i could figure out if i had a better
understanding of coordinate transforms and camera math.
glTF defines Y-up right-handed coordinates and CCW front face, and to my
knowledge it should match vulkan just fine after a 180deg rotation around
the X axis, but it doesn't.