bring back the hacky tonemapping
did a bit more research on it. was trying to calculate the average
luminance with subgroups in the fragment shader, but subgroups have
seams between triangles, so it produces wrong results. So pretty much
it will need an extra rgba16f render buffer and extra passes. Will leave
it for later.